Renamed ActionDescriptorImpl to SpringActionDescriptor and moved to spring subpackage
