+++ src/test/java/de/escalon/hypermedia/spring/uber/UberUtilsTest.java
@@ -11,8 +11,7 @@
 package de.escalon.hypermedia.spring.uber;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
-import de.escalon.hypermedia.affordance.ActionDescriptor;
-import de.escalon.hypermedia.affordance.ActionDescriptorImpl;
+import de.escalon.hypermedia.spring.SpringActionDescriptor;
 import org.junit.Test;
 import org.springframework.hateoas.Link;
 import org.springframework.hateoas.Resource;
@@ -36,7 +35,7 @@ public class UberUtilsTest {
 
     @Test
     public void linkGetToUberNode() throws Exception {
-        UberNode linkNode = UberUtils.toUberLink("/foo", new ActionDescriptorImpl("get", RequestMethod.GET.name()), Link.REL_SELF);
+        UberNode linkNode = UberUtils.toUberLink("/foo", new SpringActionDescriptor("get", RequestMethod.GET.name()), Link.REL_SELF);
         assertEquals(Arrays.asList(Link.REL_SELF), linkNode.getRel());
         assertEquals("/foo", linkNode.getUrl());
         assertNull(linkNode.getModel());
@@ -46,7 +45,7 @@ public class UberUtilsTest {
     @Test
     public void linkPostToUberNode() throws Exception {
         // TODO create a Link with variables separate from URITemplate for POST
-        UberNode linkNode = UberUtils.toUberLink("/foo{?foo,bar}", new ActionDescriptorImpl("post", RequestMethod.POST.name()), Link.REL_SELF);
+        UberNode linkNode = UberUtils.toUberLink("/foo{?foo,bar}", new SpringActionDescriptor("post", RequestMethod.POST.name()), Link.REL_SELF);
         assertEquals(Arrays.asList(Link.REL_SELF), linkNode.getRel());
         assertEquals("/foo", linkNode.getUrl());
         assertEquals("foo={foo}&bar={bar}", linkNode.getModel());
